Redid the tornado particle runner

+++ src/main/java/net/dohaw/blackclover/util/MathHelper.java
@@ -2,7 +2,7 @@ package net.dohaw.blackclover.util;
 
 public class MathHelper {
 
-    public static double angle(int i, int pointsInCircle){
+    public static double angle(double i, int pointsInCircle){
         return 2 * Math.PI * i / pointsInCircle;
     }
